update firmware path for linux 6.9 without kernel patch
the alternative to this is rebasing the wifi patch on to 6.9 or swapping the old kernel patch for valve's hack - renaming the directory to QCA206X and changing the firmware directory to that. this is marked as not-for-upstream because it will break current devices.

maybe we have this done as a part of the driver manager instead until the upstream firmware is fixed.